Refactor and fix ui tests
